requirements.txt
async-timeout==3.0.1
discord.py==1.7.2
requests==2.25.1
urllib3==1.26.5
websocket-client==1.0.1
pathspec==0.8.1
asyncpraw==7.6.1
asyncprawcore==2.3.0
python-dotenv==0.21.1
